+++ integration-tests/grpc-interceptors/src/main/java/io/quarkus/grpc/examples/interceptors/HeaderClientInterceptor.java
@@ -21,6 +21,8 @@ public class HeaderClientInterceptor implements ClientInterceptor {
 
     private static final Logger logger = Logger.getLogger(HeaderClientInterceptor.class.getName());
 
+    static volatile boolean invoked = false;
+
     static final Metadata.Key<String> CUSTOM_HEADER_KEY = Metadata.Key.of("custom_client_header_key",
             Metadata.ASCII_STRING_MARSHALLER);
 
@@ -43,6 +45,7 @@ public class HeaderClientInterceptor implements ClientInterceptor {
                                 // you can use {@link io.grpc.stub.MetadataUtils#attachHeaders}
                                 // directly to send header
                                 //
+                                invoked = true;
                                 logger.info("header received from server:" + headers);
                                 super.onHeaders(headers);
                             }
