+++ integration-tests/grpc-interceptors/src/main/java/io/quarkus/grpc/examples/interceptors/HelloWorldEndpoint.java
@@ -4,6 +4,7 @@ import javax.inject.Inject;
 import javax.ws.rs.GET;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
+import javax.ws.rs.core.Response;
 
 import examples.GreeterGrpc;
 import examples.HelloReply;
@@ -24,8 +25,11 @@ public class HelloWorldEndpoint {
 
     @GET
     @Path("/blocking/{name}")
-    public String helloBlocking(@PathParam("name") String name) {
-        return blockingHelloService.sayHello(HelloRequest.newBuilder().setName(name).build()).getMessage();
+    public Response helloBlocking(@PathParam("name") String name) {
+        HeaderClientInterceptor.invoked = false;
+        HelloReply helloReply = blockingHelloService.sayHello(HelloRequest.newBuilder().setName(name).build());
+
+        return Response.ok(helloReply.getMessage()).header("intercepted", HeaderClientInterceptor.invoked).build();
     }
 
     @GET
